==> analysis/clustering/01_nn_clustering_kd6_essential.Rmd <==
---
title: "NN Clustering on KD6 Essential Data"
author: "Mervin M Fansler"
date: "`r format(Sys.time(), '%d %B %Y')`"
output: 
  html_document: 
    code_folding: show
    toc: true
    toc_float: true
---

# Purpose

We perform clustering of features and perturbations to identify isoforms that exhibit
similar perturbation outcomes.

# Initialization

## Libraries
```{r libs, message=FALSE, warning=FALSE}
library(magrittr)
library(tidyverse)
library(pheatmap)
library(bluster)
```

## Parameters
```{r set_params}
set.seed(20210818)

## filtering options
## number of cells in perturbation
MIN_CELLS=30
## max NTPs with NA WUI
MAX_NTP_NA=0.20
## min avg TPM in NTPs
MIN_MEAN_TPM=20

## input files
FILE_SE_WUI="tbl/df_wui_kd6_essential_ui10.Rds"

## output files
FILE_OUT_TARGETS_RAW="tbl/df_target_raw_nnclusters_kd6_essential_ui10.csv"
FILE_OUT_TARGETS="tbl/df_target_nnclusters_kd6_essential_ui10.csv"
FILE_OUT_GENES="tbl/df_gene_nnclusters_kd6_essential_ui10.csv"
FILE_OUT_FINAL="img/heatmap-kd6-essential-nnclusters-labeled.pdf"

RDS_DWUI=sprintf("data/dwui/%s-dwui-target-gene-kd6-essential.Rds", format(Sys.time(), '%Y%m%d'))
RDS_ZDWUI=sprintf("data/dwui/%s-zdwui-target-gene-kd6-essential.Rds", format(Sys.time(), '%Y%m%d'))

## aesthetics
NCOLORS=100
COLORS_BWR <- colorRampPalette(c("blue", "white", "red"))(NCOLORS)
COLORS_MKY <- colorRampPalette(c("magenta", "black", "yellow"))(NCOLORS)
COLORS_YKM <- colorRampPalette(c("yellow", "black", "magenta"))(NCOLORS)
COLORS_RKG <- colorRampPalette(c("red", "black", "green"))(NCOLORS)
breaks_dwui <- seq(-0.5, 0.5, length.out=NCOLORS + 1)
breaks_zdwui <- seq(-4, 4, length.out=NCOLORS + 1)
breaks_zdwui_broad <- seq(-6, 6, length.out=NCOLORS + 1)
breaks_pca <- seq(-20, 20, length.out=NCOLORS + 1)
```

# Data
## Loading
```{r load_data, message=FALSE}
df_wui <- readRDS(FILE_SE_WUI)
```

## Preprocessing

```{r prepare_data}
sgid2gene <- df_wui %>%
    dplyr::select(sgID_AB, target_gene) %>%
    distinct(sgID_AB, target_gene) %>%
    deframe()

ens2gene <- df_wui %>%
    dplyr::select(gene_id, gene_name) %>%
    distinct(gene_id, gene_name) %>%
    deframe()

convert_rownames <- function (mat, in2out) {
    mat %>% set_rownames(in2out[rownames(.)])
}

convert_colnames <- function (mat, in2out) {
    mat %>% set_colnames(in2out[colnames(.)])
}

df_ntp <- filter(df_wui, target_gene == 'non-targeting') %>%
    group_by(gene_id, gene_name) %>%
    filter(mean(is.na(wui)) <= MAX_NTP_NA) %>%
    filter(mean(tpm) >= MIN_MEAN_TPM) %>%
    filter(!is.na(wui)) %>%
    summarize(mean_wui=weighted.mean(wui, n_cells), .groups='drop',
              sd_wui=sqrt(sum((wui-mean_wui)^2)/n()))

df_dwui <- df_wui %>%
    filter(target_gene != 'non-targeting',
           n_cells >= MIN_CELLS) %>%
    inner_join(df_ntp, by=c("gene_id", "gene_name")) %>%
    mutate(dwui=wui-mean_wui) %>%
    dplyr::select(gene_id, sgID_AB, wui, dwui)

wui_gene_target <- df_dwui %>%
    dplyr::select(gene_id, sgID_AB, wui) %>%
    pivot_wider(id_cols="gene_id", names_from="sgID_AB", values_from="wui") %>%
    column_to_rownames("gene_id") %>%
    as.matrix

dwui_gene_target <- df_dwui %>%
    dplyr::select(gene_id, sgID_AB, dwui) %>%
    pivot_wider(id_cols="gene_id", names_from="sgID_AB", values_from="dwui") %>%
    column_to_rownames("gene_id") %>%
    as.matrix 

zdwui_gene_target <- t(dwui_gene_target) %>% scale(center=FALSE) %>% t()

zdwui_gene_target_clean <- zdwui_gene_target
zdwui_gene_target_clean[is.na(zdwui_gene_target_clean)] <- 0
```

# PCA
## Compute PCA 
```{r pc_genes}
res_pca <- BiocSingular::runPCA(t(zdwui_gene_target_clean), rank=30, center=FALSE)
mat_pca <- res_pca$x
```

## Heatmap - All Components
```{r plt_pca30, fig.width=4, fig.height=8}
pheatmap(mat_pca, 
         color=COLORS_YKM,
         breaks=breaks_pca, 
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```
### Compute Clusters
```{r compute_target_clusters1}
clusters_targets_raw <- clusterRows(mat_pca, NNGraphParam(k=5, cluster.fun="walktrap"))
names(clusters_targets_raw) <- rownames(mat_pca)

idx_clusters_targets <- order(clusters_targets_raw)

df_row_annots <- data.frame(cluster_id_target=clusters_targets_raw, 
                            row.names=names(clusters_targets_raw))

table(clusters_targets_raw)
```

### Export Target Clusters
```{r export_target_clusters}
df_wui %>%
    distinct(target_gene, target_gene_id, sgID_AB) %>%
    inner_join(tibble(sgID_AB=rownames(mat_pca), cluster_id=clusters_targets_raw), by="sgID_AB") %>%
    write_csv(FILE_OUT_TARGETS_RAW)
```

### Plot Clusters
```{r plt_pca_wt_clusters_raw, fig.width=4, fig.height=8}
pheatmap(mat_pca[idx_clusters_targets,], 
         color=COLORS_YKM,
         breaks=breaks_pca, 
         fontsize_col=1, fontsize_row=1,
         annotation_row=df_row_annots,
         show_colnames=FALSE, show_rownames=FALSE,
         cluster_rows=FALSE, cluster_cols=TRUE, scale='none',
         cutree_rows=NA)
```

It appears clusters 8 and 15 represent sets of perturbations with null responses.

## Z-DWUI Plot
### Potential Null Response Clusters
We want to check the potential null-response clusters in the DWUI representation.

#### Cluster 8
```{r plt_noise_target_cluster8_zdwui, fig.width=8, fig.height=8}
idx_cluster <- clusters_targets_raw %>% `[`(. == 8) %>% names()
pheatmap(t(zdwui_gene_target[,idx_cluster]), 
         color=COLORS_YKM,
         breaks=breaks_zdwui,
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```

#### Cluster 15
```{r plt_noise_target_cluster15_zdwui, fig.width=8, fig.height=8}
idx_cluster <- clusters_targets_raw %>% `[`(. == 15) %>% names()
pheatmap(t(zdwui_gene_target[,idx_cluster]), 
         color=COLORS_YKM,
         breaks=breaks_zdwui,
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```

This looks mostly like noise in both clusters.

### All Other Clusters

```{r plt_signal_target_clusters_zdwui, fig.width=8, fig.height=8}
idx_cluster <- clusters_targets_raw %>% `[`(!(. %in%  c(8,15))) %>% names
pheatmap(t(zdwui_gene_target[,idx_cluster]), 
         color=COLORS_YKM,
         breaks=breaks_zdwui,
         fontsize_col=1, fontsize_row=1,
         annotation_row=df_row_annots,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```

This looks good. Now we need to eliminate the noise columns (readout genes).

## Establish Regulating Perturbations
```{r select_signal_targets}
idx_signal_targets <- clusters_targets_raw %>% `[`(!(. %in%  c(8,15))) %>% names
```

We identify `r length(idx_signal_targets)` genes as coherent response genes.

# Cluster Response Genes
## Reduced PCA
```{r pc_targets}
res_pca2 <- BiocSingular::runPCA(zdwui_gene_target_clean[,idx_signal_targets], rank=30, center=FALSE)
mat_pca2 <- res_pca2$x
```

```{r plt_pca_reduced_all, fig.width=8, fig.height=4}
pheatmap(t(mat_pca2), 
         color=COLORS_YKM,
         breaks=breaks_pca, 
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```

### Compute Gene Clusters
```{r compute_gene_clusters}
clusters_genes <- clusterRows(mat_pca2, NNGraphParam(k=4, cluster.fun="walktrap"))
names(clusters_genes) <- rownames(mat_pca2)

idx_clusters_genes <- order(clusters_genes)

df_col_annots <- data.frame(cluster_id_gene=clusters_genes, 
                            row.names=names(clusters_genes))

table(clusters_genes)
```

### Export Gene Clusters
```{r export_gene_clusters}
df_wui %>%
    distinct(gene_name, gene_id) %>%
    inner_join(enframe(clusters_genes, name="gene_id", value="cluster_id"), by="gene_id") %>%
    write_csv(FILE_OUT_GENES)
```

### Plot Clusters
```{r plt_pca_wt_clusters_genes, fig.width=8, fig.height=4}
pheatmap(t(mat_pca2[idx_clusters_genes,]), 
         color=COLORS_YKM,
         breaks=breaks_pca, 
         fontsize_col=1, fontsize_row=1,
         annotation_col=df_col_annots,
         show_colnames=FALSE, show_rownames=FALSE,
         cluster_rows=TRUE, cluster_cols=FALSE, scale='none')
```

#### Cluster 9
```{r plt_noise_gene_cluster9_zdwui, fig.width=4, fig.height=4}
idx_cluster <- clusters_genes %>% `[`(. == 9) %>% names()
pheatmap(t(zdwui_gene_target[idx_cluster, idx_signal_targets]), 
         color=COLORS_YKM,
         breaks=breaks_zdwui,
         annotation_row=df_row_annots,
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```


#### Cluster 12
```{r plt_noise_gene_cluster12_zdwui, fig.width=4, fig.height=4}
idx_cluster <- clusters_genes %>% `[`(. == 12) %>% names()
pheatmap(t(zdwui_gene_target[idx_cluster, idx_signal_targets]), 
         color=COLORS_YKM,
         breaks=breaks_zdwui,
         annotation_row=df_row_annots,
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```

#### Cluster 13
```{r plt_noise_gene_cluster13_zdwui, fig.width=4, fig.height=4}
idx_cluster <- clusters_genes %>% `[`(. == 13) %>% names()
pheatmap(t(zdwui_gene_target[idx_cluster, idx_signal_targets]), 
         color=COLORS_YKM,
         breaks=breaks_zdwui,
         annotation_row=df_row_annots,
         fontsize_col=1, fontsize_row=1,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE)
```

Clusters 9, 12, and 13 appear to be mostly non-responsive.

## Establish Signal Genes
```{r select_signal_genes}
idx_signal_genes <- clusters_genes %>% `[`(!(. %in%  c(9,12,13))) %>% names
```

We identify `r length(idx_signal_genes)` genes as coherent response genes.

```{r plt_signal_gene_clusters_zdwui, fig.width=8, fig.height=8}
pheatmap(t(zdwui_gene_target[idx_signal_genes, idx_signal_targets]), 
         color=COLORS_BWR,
         breaks=breaks_zdwui,
         fontsize_col=1, fontsize_row=1,
         annotation_row=df_row_annots, 
         annotation_col=df_col_annots,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=TRUE, cluster_cols=TRUE, clustering_method="complete")
```


# Recompute Target Clusters
```{r compute_target_clusters_final}
zdwui_target_gene_final <- t(zdwui_gene_target_clean[idx_signal_genes, idx_signal_targets])
res_pca3 <- BiocSingular::runPCA(zdwui_target_gene_final, rank=30, center=FALSE)
mat_pca3 <- res_pca3$x

clusters_targets_final <- clusterRows(mat_pca3, BLUSPARAM=NNGraphParam(k=3, cluster.fun="walktrap"))
names(clusters_targets_final) <- rownames(mat_pca3)

clusters_genes_final <- clusters_genes %>% `[`(!(. %in% c(9,12,13)))

idx_clusters_targets_final <- clusters_targets_final %>% sort %>% names
idx_clusters_genes_final <- clusters_genes_final %>% sort %>% names

df_row_annots_final <- data.frame(cluster_id_target=clusters_targets_final, 
                                  row.names=names(clusters_targets_final))

table(clusters_targets_final)
```

### Export Final Target Clusters
```{r export_final_target_clusters}
df_wui %>%
    distinct(target_gene, target_gene_id, sgID_AB) %>%
    inner_join(tibble(sgID_AB=rownames(mat_pca3), cluster_id=clusters_targets_final), by="sgID_AB") %>%
    write_csv(FILE_OUT_TARGETS)
```

# Final Heatmap
```{r plt_final_clusters_zdwui, fig.width=8, fig.height=8}
pheatmap(zdwui_target_gene_final[idx_clusters_targets_final, idx_clusters_genes_final], 
         color=COLORS_BWR,
         breaks=breaks_zdwui,
         fontsize_col=1, fontsize_row=1,
         annotation_row=df_row_annots_final, 
         annotation_col=df_col_annots,
         show_colnames=FALSE, show_rownames=FALSE, scale='none', 
         cluster_rows=FALSE, cluster_cols=FALSE, clustering_method="complete",
         gaps_row=cumsum(table(clusters_targets_final)),
         gaps_col=cumsum(table(clusters_genes_final)) %>% `[`(!duplicated(.)),
         cutree_rows=NA, cutree_cols=NA)
```


# Final Joint Plot
```{r joint_plot, figure.width=16, fig.height=16}
pheatmap(zdwui_target_gene_final[idx_clusters_targets_final, idx_clusters_genes_final], 
         color=COLORS_BWR,
         breaks=breaks_zdwui,
         fontsize_col=1, fontsize_row=1,
         annotation_row=df_row_annots_final, 
         annotation_col=df_col_annots,
         show_colnames=TRUE, show_rownames=TRUE, scale='none', 
         labels_row=sgid2gene[idx_clusters_targets_final],
         labels_col=ens2gene[idx_clusters_genes_final],
         cluster_rows=FALSE, cluster_cols=FALSE, clustering_method="complete",
         gaps_row=cumsum(table(clusters_targets_final)),
         gaps_col=cumsum(table(clusters_genes_final)) %>% `[`(!duplicated(.)),
         cutree_rows=NA, cutree_cols=NA,
         filename=FILE_OUT_FINAL, width=16, height=16)
```


```{r export_matrices}
saveRDS(t(dwui_gene_target), RDS_DWUI)
saveRDS(t(zdwui_gene_target), RDS_ZDWUI)
```

---

# Runtime Details
## Session Info
```{r sesh_info, echo=FALSE}
sessionInfo()
```

## Conda Environment
```{bash comment="", echo=FALSE}
if ! command -v conda &> /dev/null
then
  echo "Conda not detected."
elif [ -z "${CONDA_PREFIX}" ]
then
  echo "No active Conda environment."
else
  echo "## Conda Environment YAML"
  conda env export
fi
```
